pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "agent-bench"
version = "0.1.0"
description = "Benchmark suite for evaluating how AI-agent-native a website is"
readme = "README.md"
requires-python = ">=3.11"
license = "MIT"
dependencies = [
    "click>=8.1",
    "pydantic>=2.0",
    "httpx>=0.27",
    "beautifulsoup4>=4.12",
    "pyyaml>=6.0",
    "rich>=13.0",
]

[project.optional-dependencies]
browser = [
    "playwright>=1.40",
    "browser-use>=0.1",
]
dev = [
    "pytest>=8.0",
    "ruff>=0.4",
]

[project.scripts]
agent-bench = "agent_bench.cli:cli"

[tool.hatch.build.targets.wheel]
packages = ["src/agent_bench"]

[tool.ruff]
target-version = "py311"
src = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]
